moved es related code to ElasticSearchService
